// src/app/models/movie.ts
export class Movie {
    adult: Boolean;
    backdrop_path: String;
    belongs_to_collection: Object;  
    budget: Number;
    genres: Object[];
    homepage: String;
    id: Number;
    imdb_id: String;
    original_language: String;
    original_title: String;
    overview: String;
    popularity: Number;
    poster_path: String;
    production_companies: Object[];
    production_countries: Object[];
    release_date: String;
    revenue: Number;
    runtime: Number;
    spoken_languages: Object[];
    status: String;
    tagline: String;
    title: String;
    video: Boolean;
    vote_average: Number;
    vote_count: Number; 
}